backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
python-multipart
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
